+++ webpack.config.js
@@ -4,7 +4,7 @@
 const path = require("path");
 
 /** @type {import('webpack').Configuration} */
-const config = {
+const extensionConfig = {
   target: "node",
   mode: "none", // set by npm scripts via --mode when building
   entry: "./src/extension.ts",
@@ -19,12 +19,12 @@ const config = {
     vscode: "commonjs vscode"
   },
   resolve: {
-    extensions: [".ts", ".js"],
+    extensions: [".ts", ".tsx", ".js", ".jsx"],
   },
   module: {
     rules: [
       {
-        test: /\.(ts)$/,
+        test: /\.(ts|tsx)$/,
         exclude: [/node_modules\//, /src[\\\/]test[\\\/]?/],
         use: [
           {
@@ -51,4 +51,49 @@ const config = {
   }
 };
 
-module.exports = config;
+const webviewConfig = {
+  target: "web",
+  mode: "none",
+  entry: {
+    main: path.resolve(__dirname, "src", "webview", "main.tsx")
+  },
+  output: {
+    path: path.resolve(__dirname, "out", "webview"),
+    filename: "[name].js",
+    devtoolModuleFilenameTemplate: "../../[resource-path]"
+  },
+  devtool: "source-map",
+  resolve: {
+    extensions: [".ts", ".tsx", ".js", ".jsx"],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(ts|tsx)$/,
+        exclude: /node_modules\//,
+        use: [
+          {
+            loader: "ts-loader",
+            options: {
+              transpileOnly: false
+            }
+          }
+        ]
+      },
+      {
+        test: /\.js$/,
+        enforce: "pre",
+        use: [
+          {
+            loader: "source-map-loader"
+          }
+        ]
+      }
+    ]
+  },
+  infrastructureLogging: {
+    level: "warn",
+  }
+};
+
+module.exports = [extensionConfig, webviewConfig];